Corrigindo o tipo do valor na função validar do Tempo

+++ dominios/codigos/tempo.cpp
@@ -1,7 +1,7 @@
 #include "../headers/tempo.hpp"
 
 
-bool Tempo::validar(uint8_t valor){
+bool Tempo::validar(int valor){
     if (valor < 1 or valor > 365)
         return false;
     return true;
